Refactor: Align CMake, code structure, and CI with best practices.
This commit applies a series of changes to improve your project's
CMake configuration, code organization, and GitHub Actions CI setup,
aiming for better compliance with JUCE and CMake best practices.

Key changes include:

1.  **Code Structure:**
    *   I relocated `PresetManager.h` and `PresetManager.cpp` from
        `plugin/.../Data/` to `plugin/include/Pointilsynth/` and
        `plugin/source/` respectively, removing the `Data` subdirectories.
        This centralizes plugin code as per the documented structure.

2.  **Plugin CMake (`plugin/CMakeLists.txt`):**
    *   I set `IS_SYNTH` to `TRUE` for the JUCE plugin definition.
    *   I expanded `HEADER_FILES` to include all public headers in
        `plugin/include/Pointilsynth/`, improving IDE integration.
    *   I uncommented `source_group(TREE ...)` for better source file
        organization in IDEs.
    *   I corrected `INCLUDE_DIR` variable typo from `PointilSynth` to `Pointilsynth`.
    *   I expanded plugin `FORMATS` to `Standalone VST3 AU AAX`.
    *   I removed erroneous `GTest::gtest` linkage from the plugin target.
    *   I cleaned up a redundant `nlohmann_json::nlohmann_json` linkage.

3.  **GitHub Actions CI:**
    *   **Pre-merge (`.github/workflows/cmake.yml`):**
        *   I corrected `actions/cache` path to `${{github.workspace}}/libs`.
        *   I added `Release` build type to the Linux matrix, so both Debug
            and Release configurations are built and tested on pushes/PRs
            to `main`.
    *   **Post-merge (`.github/workflows/post-merge-builds.yml`):**
        *   I created a new workflow for builds that run only after merging to `main`.
        *   This workflow builds and tests `Release` configurations on
            `macos-latest` and `windows-latest`.

**Build and Test Verification:**
My attempt to perform a final build and test cycle
failed due to an execution environment issue that prevented CMake from running.
Therefore, the primary validation of these changes will rely on the
updated GitHub Actions workflows.
