+++ plugin/include/Pointilsynth/PresetManager.h
